+++ NEWS.md
@@ -1,6 +1,9 @@
 #Update
-The Firebrowse API changed to version 1.1.2 beta. And some changes took place
+Fixed some minor bugs when downloading CSV and some bigger issues with TSV
+Also updated API version 1.1.21.
 
-## Metadata.Counts
-This method has been fixed, it was broken when specifying the `data_type` paramter
+## Samples
+Removed ClinicalTier1 from Samples and added Clinical & Clinical_FH instead.
 
+## Metadata
+Here also clinicalTier1 got replaced by Clinical & Clinical_FH